requirements.txt
# Lung Cancer Detection - DCA-Net Project Dependencies

# ============== Core ==============
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ============== Medical Imaging ==============
SimpleITK>=2.2.0
pydicom>=2.4.0
nibabel>=5.0.0
scikit-image>=0.21.0

# ============== Deep Learning ==============
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0
albumentations>=1.3.0

# ============== Visualization ==============
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# ============== Experiment Tracking ==============
wandb>=0.15.0
tensorboard>=2.13.0

# ============== Evaluation ==============
scikit-learn>=1.3.0

# ============== Model Optimization ==============
onnx>=1.14.0
onnxruntime>=1.15.0

# ============== Utilities ==============
tqdm>=4.65.0
pyyaml>=6.0
click>=8.1.0
kagglehub
